added the splash pictures
